# Conflicts:
#	.gitignore
#	build.gradle
#	common/src/main/java/net/sistr/littlemaidmodelloader/maidmodel/ModelLittleMaidBase.java
#	common/src/main/java/net/sistr/littlemaidmodelloader/maidmodel/ModelRenderer.java
#	common/src/main/java/net/sistr/littlemaidmodelloader/maidmodel/compat/GLCompat.java
#	common/src/main/java/net/sistr/littlemaidmodelloader/resource/classloader/MultiModelClassTransformer.java
#	common/src/main/java/net/sistr/littlemaidmodelloader/resource/manager/LMTextureManager.java
#	common/src/main/resources/littlemaidmodelloader-common.mixins.json
#	fabric/build.gradle
#	fabric/src/main/resources/fabric.mod.json
#	forge/src/main/java/net/sistr/littlemaidmodelloader/entrypoint/ModEntryPoint.java
#	gradle.properties

+++ common/src/main/java/net/sistr/littlemaidmodelloader/maidmodel/ModelBase.java
@@ -83,7 +83,7 @@ public abstract class ModelBase extends AbstractModelBase {
     }
 
     public static float mh_sqrt(double d) {
-        return MathHelper.sqrt((float) d);
+        return MathHelper.sqrt(d);
     }
 
     public static int mh_floor(float f) {
